refactor: move pass topo resolver out of main to reduce merge conflicts
